requirements.txt
# Core LangChain dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# Model providers
ollama>=0.4.0
langchain-openai>=0.2.0

# Tools and utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0

# Data handling
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Additional utilities
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.0

# Optional but recommended
colorama>=0.4.6  # Colored terminal output
tqdm>=4.66.0     # Progress bars
